move mtric to component folder, too
